move Handle to common-types
