+++ services/brig/src/Brig/User/Search/Index/Types.hs
@@ -5,6 +5,7 @@ module Brig.User.Search.Index.Types where
 import Brig.Types.User
 import Control.Lens (makeLenses)
 import Data.Aeson
+import Data.Handle (Handle)
 import Data.Id
 import Data.Json.Util ((#))
 import Database.V5.Bloodhound hiding (key)
