+++ services/brig/test/integration/API/Search/Util.hs
@@ -6,6 +6,7 @@ import Brig.Types
 import Control.Monad.Catch (MonadCatch)
 import Control.Monad.Fail (MonadFail)
 import Data.Aeson (decode, encode)
+import Data.Handle (Handle (Handle))
 import Data.Id
 import Data.Text.Encoding (encodeUtf8)
 import Imports
